\name{BasicFun}
\alias{BasicFun}
\title{Basic Data Information}
\description{
Basic Data Information in individuals (abundances) data.
}
\usage{
BasicFun(x1, x2, B)
}

\arguments{
  \item{x1}{the species abundances for community 1}
  \item{x2}{the species abundances for community 2}
  \item{B}{number of bootstrap resampling times}
}