\name{BasicFun.Sam}
\alias{BasicFun.Sam}
\title{Basic Data Information}
\description{
Basic Data Information in sample-base (incidence) data.
}
\usage{
BasicFun.Sam(y1, y2, B)
}
\arguments{
  \item{y1}{the first entry should be total number of sampling units, and followed by species incidence frequencies in community 1}
  \item{y2}{the first entry should be total number of sampling units, and followed by species incidence frequencies in community 2}
  \item{B}{the number of bootstrap resampling times, default is 200.}
}
