\name{ChaoShared-package}
\alias{ChaoShared-package}
\docType{package}
\title{
Statistical package "\pkg{ChaoShared}"
}
\description{
The purpose of this R package \pkg{ChaoShared} estimates for the number 
  of shared species in two communities based on two sampling schemes.
}
\details{
\tabular{ll}{
Package: \tab ChaoShared\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-01-18\cr
License: \tab GPL (>= 2)\cr
URL: \tab \url{http://chao.stat.nthu.edu.tw/blog/}
}
functions : \kbd{ChaoShared}
}
\author{
Y.H. Lee ; K.S. Tseng ; Anne Chao \cr
Maintainer : Anne Chao \email{chao@stat.nthu.edu.tw}
}
\seealso{
   \code{\link[ChaoShared]{ChaoShared}}
}
\keyword{shared species}
