\name{print.BasicFun}
\alias{print.BasicFun}
\title{Print Method for Basic data information}
\description{
Print method for basic data information in individuals (abundances) data.
}

\usage{
\method{print}{BasicFun}(x, \dots)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{a \code{\link{BasicFun}} objext computed by BasicFun}
  \item{...}{additional objects of the same type.}
}

