\name{print.BasicFun.Sam}
\alias{print.BasicFun.Sam}
\title{Print Method for Basic data information}
\description{
Print method for basic data information in sample-base (incidence) data.
}
\usage{
\method{print}{BasicFun.Sam}(x, \dots)
}
\arguments{
  \item{x}{a \code{\link{BasicFun.Sam}} objext computed by BasicFun.Sam}
  \item{...}{additional objects of the same type.}
}
